move time filter to templates
